Housekeeping - split up code more logically and rm unnecessary .es6 ext
