+++ javascripts/unformatted-code-detector/core/strip-ignored-content.js
