+++ test/unit/from-bug-reports-test.js
@@ -1,7 +1,7 @@
 import { module } from "qunit";
 import { expectAll } from "../helpers/utils";
 import { falseNegatives, falsePositives } from "../fixtures/bug-reports";
-import { detectUnformattedCode } from "../unformatted_code_detector/lib/detect-code";
+import { detectUnformattedCode } from "../unformatted-code-detector/core/detect-code";
 
 module("bug reports", () => {
   expectAll(detectUnformattedCode, falseNegatives, true);
