+++ test/unit/sequential-lines-test.js
@@ -2,11 +2,11 @@ import { module, test } from "qunit";
 import {
   getCodeEnergy,
   numSequentialLinesWithThresholdCodeEnergy,
-} from "../unformatted_code_detector/lib/detect-code";
+} from "../unformatted-code-detector/core/detect-code";
 import {
   codeEnergyValues,
   CodeEnergyLevels,
-} from "../unformatted_code_detector/lib/code-energy";
+} from "../unformatted-code-detector/core/code-energy";
 
 module("sequential lines", ({ beforeEach }) => {
   beforeEach(() => {
